Unify command mapping in the functionality tests
