+++ tests/functionality_all_platforms.py
@@ -7,7 +7,6 @@
 import sys
 import copy
 import inspect
-import platform
 import unittest
 import functools
 
@@ -29,7 +28,7 @@ from components.scmprovider import SCMProvider
 from apis.taskcluster import TaskclusterProvider
 from apis.phabricator import PhabricatorProvider
 
-from tests.functionality_utilities import TRY_OUTPUT, ARC_OUTPUT, CONDUIT_USERNAME_SEARCH_OUTPUT, CONDUIT_EDIT_OUTPUT, GIT_DIFF_FILES_CHANGES, GIT_COMMIT_BODY
+from tests.functionality_utilities import AssertFalse, SHARED_COMMAND_MAPPINGS, TRY_OUTPUT, CONDUIT_EDIT_OUTPUT
 from tests.mock_commandprovider import TestCommandProvider
 from tests.mock_libraryprovider import MockLibraryProvider
 from tests.mock_treeherder_server import MockTreeherderServer, reset_seen_counters
@@ -52,38 +51,12 @@ def DEFAULT_EXPECTED_VALUES(git_pretty_output_func, get_filed_bug_id_func):
     })
 
 
-def AssertFalse(a=False, b=False, c=False):
-    assert False, "We should not have called this function in this test."
-
-
 def COMMAND_MAPPINGS(expected_values, abandon_callback):
-    def echo_str(s):
-        if platform.system() != "Windows":
-            return s.replace("echo {", "echo '{")
-        return s
-
-    return {
-        "./mach vendor": lambda: expected_values.library_new_version_id() + " 2020-08-21T15:13:49.000+02:00",
+    ret = SHARED_COMMAND_MAPPINGS(expected_values, abandon_callback)
+    ret.update({
         "./mach try auto": lambda: TRY_OUTPUT(expected_values.try_revision_id()),
-        "hg commit": lambda: "",
-        "hg checkout -C .": lambda: "",
-        "hg purge .": lambda: "",
-        "hg status": lambda: "",
-        "hg strip": lambda: "",
-        "arc diff --verbatim": lambda: ARC_OUTPUT % (expected_values.phab_revision_func(), expected_values.phab_revision_func()),
-        echo_str("echo {\"constraints\""): lambda: CONDUIT_USERNAME_SEARCH_OUTPUT,
-        echo_str("echo {\"transactions\": [{\"type\":\"reviewers.set\""): lambda: CONDUIT_EDIT_OUTPUT,
-        echo_str("echo {\"transactions\": [{\"type\":\"abandon\""): abandon_callback if abandon_callback else AssertFalse,
-        echo_str("echo {\"transactions\": [{\"type\":\"bugzilla.bug-id\""): lambda: CONDUIT_EDIT_OUTPUT,
-        "git log -1 --oneline": lambda: "0481f1c (HEAD -> issue-115-add-revision-to-log, origin/issue-115-add-revision-to-log) Issue #115 - Add revision of updatebot to log output",
-        "git clone https://example.invalid .": lambda: "",
-        "git merge-base": lambda: "_current",
-        "git diff --name-status": lambda: GIT_DIFF_FILES_CHANGES,
-        "git log --pretty=%H|%ai|%ci": lambda cmd: "\n".join(expected_values.git_pretty_output_func("_current" not in cmd)),
-        "git log --pretty=%s": lambda: "Roll SPIRV-Tools from a61d07a72763 to 1cda495274bb (1 revision)",
-        "git log --pretty=%an": lambda: "Tom Ritter",
-        "git log --pretty=%b": lambda: GIT_COMMIT_BODY,
-    }
+    })
+    return ret
 
 
 ALL_BUGS = False
